feat: モビルスーツ更新処理でバージョン番号を確認するように修正
refs issue-012

+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/application/usecase/mobilesuit/UpdateMobileSuitUseCase.java
@@ -8,6 +8,7 @@ import jp.co.sunarch.mobilesuitDatabase.application.command.mobilesuit.UpdateMob
 import jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.MobileSuitQueryService;
 import jp.co.sunarch.mobilesuitDatabase.application.service.mobilesuit.MobileSuitRecodeService;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.MobileSuit;
+import jp.co.sunarch.mobilesuitDatabase.util.exception.MobileSuitDataBaseConflictException;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -19,11 +20,14 @@ public class UpdateMobileSuitUseCase {
 
 	public void execute(UpdateMobileSuitCommand command) {
 		// 更新対象のMobileSuitを取得
-		MobileSuit mobileSuit = mobileSuitQueryService.getMobileSuitById(command.getMsId());
+		MobileSuit before = mobileSuitQueryService.getMobileSuitById(command.getMsId());
 
-		// 更新対象のMobileSuitに更新項目をセット
+		// 更新用のMobileSuitを作成
+		MobileSuit mobileSuit = new MobileSuit();
+		mobileSuit.setMsId(command.getMsId());
 		mobileSuit.setModelNumber(command.getModelNumber());
 		mobileSuit.setMsName(command.getMsName());
+		mobileSuit.setMsUrl(before.getMsUrl());
 		mobileSuit.setHeadHeight(command.getHeadHeight());
 		mobileSuit.setOverallHeight(command.getOverallHeight());
 		mobileSuit.setWeight(command.getWeight());
@@ -35,8 +39,13 @@ public class UpdateMobileSuitUseCase {
 		mobileSuit.setTotalThrustersOutput(command.getTotalThrustersOutput());
 		mobileSuit.setMsOverview(command.getMsOverview());
 		mobileSuit.setAction(command.getAction());
+		mobileSuit.setInsertDate(before.getInsertDate());
 		mobileSuit.setUpdateDate(Instant.now());
-		mobileSuit.setVersion(mobileSuit.getVersion() + 1);
+		mobileSuit.setVersion(command.getVersion());
+
+		if(before.getVersion() != mobileSuit.getVersion()) {
+			throw new MobileSuitDataBaseConflictException("すでに別のユーザーによって更新された可能性があります。最新情報をご確認ください。");
+		}
 
 		if (!command.getMsMultipartFile().isEmpty()) {
 			mobileSuitRecodeService.deleteImageFile(mobileSuit.getMsUrl());
